+++ src/main/java/com/maincrud/main/service/EmployeeService.java
@@ -3,13 +3,14 @@ package com.maincrud.main.service;
 import com.maincrud.main.entity.Employee;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface EmployeeService {
     List<Employee> findAll();
 
-    Employee findById(int id);
+    Employee findById(Long id);
 
     Employee save(Employee employee);
 
-    void deleteEmployee(int id);
+    void deleteEmployee(Long id);
 }
